edit some

+++ app/Console/Commands/CronSendmail.php
@@ -8,6 +8,7 @@ use Carbon\Carbon;
 use App\Model\Pillar;
 use App\Model\FullTbp;
 use App\Model\MiniTBP;
+use App\Model\Scoring;
 use App\Helper\EmailBox;
 use App\Model\MeetingDate;
 use App\Model\EventCalendar;
@@ -106,6 +107,28 @@ class CronSendmail extends Command
             if($date->isToday() == 1){
                 $ev = Ev::where('full_tbp_id',$eventcalendar->full_tbp_id)->first();
                 ProjectMember::whereNotIn('user_id',$scoringstatuses)->delete();
+
+                $check = Scoring::where('ev_id',$ev->id)
+                                ->whereNull('user_id')
+                                ->get(); 
+                if($check->count() == 0){
+                    $userid = Scoring::where('ev_id',$ev->id)
+                            ->whereNotNull('user_id')
+                            ->first()->user_id; 
+                    $scorings = Scoring::where('ev_id',$ev->id)
+                            ->where('user_id',$userid)
+                            ->get(); 
+                    foreach ($scorings as $key => $scoring) {
+                        $new = new Scoring();
+                        $new->ev_id = $scoring->ev_id;
+                        $new->criteria_transaction_id  = $scoring->criteria_transaction_id ;
+                        $new->sub_pillar_index_id = $scoring->sub_pillar_index_id;
+                        $new->scoretype = $scoring->scoretype;
+                        $new->score = $scoring->score;
+                        $new->save();
+                    }      
+                }
+
             }
         }
     }
